🎨 format YAML-files in build process
