+++ .github/dependabot.old.yml
@@ -1,33 +1,33 @@
 # disabled in favor of renovate
 version: 2
 updates:
-  - package-ecosystem: "npm"
-    directory: "/" # Location of package manifests
-    schedule:
-      interval: "daily"
-      time: "03:00"
-      timezone: "Europe/Berlin"
-    assignees:
-      - "Suchty112"
-    commit-message:
-        prefix: "⬆️"
-        include: "scope"
-    reviewers:
-      - "Suchty112"
-      - "kdev"
-    target-branch: "dev"
-  - package-ecosystem: "npm"
-    directory: "/docs/.vuepress" # Location of package manifests
-    schedule:
-      interval: "daily"
-      time: "03:00"
-      timezone: "Europe/Berlin"
-    assignees:
-      - "Suchty112"
-    commit-message:
-        prefix: "⬆️"
-        include: "scope"
-    reviewers:
-      - "Suchty112"
-      - "kdev"
-    target-branch: "dev"
+    - package-ecosystem: 'npm'
+      directory: '/' # Location of package manifests
+      schedule:
+          interval: 'daily'
+          time: '03:00'
+          timezone: 'Europe/Berlin'
+      assignees:
+          - 'Suchty112'
+      commit-message:
+          prefix: '⬆️'
+          include: 'scope'
+      reviewers:
+          - 'Suchty112'
+          - 'kdev'
+      target-branch: 'dev'
+    - package-ecosystem: 'npm'
+      directory: '/docs/.vuepress' # Location of package manifests
+      schedule:
+          interval: 'daily'
+          time: '03:00'
+          timezone: 'Europe/Berlin'
+      assignees:
+          - 'Suchty112'
+      commit-message:
+          prefix: '⬆️'
+          include: 'scope'
+      reviewers:
+          - 'Suchty112'
+          - 'kdev'
+      target-branch: 'dev'
